% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Function.R
\name{calculate_drinking_water}
\alias{calculate_drinking_water}
\title{Calculate drinking water (vapor)}
\usage{
calculate_drinking_water(df)
}
\arguments{
\item{df}{Data frame containing DrinkingH2OIngested}
}
\value{
Numeric vector of drinking water vapor
}
\description{
Calculate drinking water (vapor)
}
\keyword{internal}
