% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Function.R
\name{calculate_drinking_water_ingested}
\alias{calculate_drinking_water_ingested}
\title{Calculate drinking water ingested}
\usage{
calculate_drinking_water_ingested(df)
}
\arguments{
\item{df}{Data frame containing water balance components}
}
\value{
Numeric vector of drinking water ingested (moles)
}
\description{
Calculate drinking water ingested
}
\keyword{internal}
