% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_dry_fecal_output}
\alias{calculate_dry_fecal_output}
\title{Calculate dry fecal output}
\usage{
calculate_dry_fecal_output(df)
}
\arguments{
\item{df}{Dataframe with FoodMassIngested and Digestibility columns}
}
\value{
Numeric vector of dry fecal output values
}
\description{
Calculate dry fecal output
}
\keyword{internal}
