% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Function.R
\name{calculate_dry_hydrogen_influx}
\alias{calculate_dry_hydrogen_influx}
\title{Calculate dry hydrogen influx}
\usage{
calculate_dry_hydrogen_influx(df)
}
\arguments{
\item{df}{Data frame containing food mass, digestibility, EEE,
and hydrogen composition}
}
\value{
Numeric vector of dry hydrogen influx
}
\description{
Calculate dry hydrogen influx
}
\keyword{internal}
