% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Function.R
\name{calculate_dry_oxygen_influx}
\alias{calculate_dry_oxygen_influx}
\title{Calculate dry oxygen influx}
\usage{
calculate_dry_oxygen_influx(df)
}
\arguments{
\item{df}{Data frame containing food mass, digestibility, EEE,
and oxygen composition}
}
\value{
Numeric vector of dry oxygen influx
}
\description{
Calculate dry oxygen influx
}
\keyword{internal}
