% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Function.R
\name{calculate_energy_density}
\alias{calculate_energy_density}
\title{Calculate energy density of food}
\usage{
calculate_energy_density(df)
}
\arguments{
\item{df}{Data frame containing food composition columns}
}
\value{
Numeric vector of energy densities (kJ/g)
}
\description{
Calculate energy density of food
}
\keyword{internal}
