% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_fecal_water_loss}
\alias{calculate_fecal_water_loss}
\title{Calculate fecal water loss}
\usage{
calculate_fecal_water_loss(df)
}
\arguments{
\item{df}{Dataframe with DryFecalOutput column}
}
\value{
Numeric vector of fecal water loss values
}
\description{
Calculate fecal water loss
}
\keyword{internal}
