% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Function.R
\name{calculate_food_mass}
\alias{calculate_food_mass}
\title{Calculate food mass ingested}
\usage{
calculate_food_mass(df)
}
\arguments{
\item{df}{Data frame containing EnergyExp, Digestibility, EEE, and
food composition}
}
\value{
Numeric vector of food mass ingested (g)
}
\description{
Calculate food mass ingested
}
\keyword{internal}
