% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Function.R
\name{calculate_free_water_in_food}
\alias{calculate_free_water_in_food}
\title{Calculate free water in food}
\usage{
calculate_free_water_in_food(df)
}
\arguments{
\item{df}{Data frame containing FoodMassIngested and freeH20food}
}
\value{
Numeric vector of free water in food (moles)
}
\description{
Calculate free water in food
}
\keyword{internal}
