% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_panting}
\alias{calculate_panting}
\title{Calculate panting water loss}
\usage{
calculate_panting(df)
}
\arguments{
\item{df}{Dataframe with WaterHeatLoss column}
}
\value{
Numeric vector of panting values
}
\description{
Calculate panting water loss
}
\keyword{internal}
