% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_urea_produced}
\alias{calculate_urea_produced}
\title{Calculate urea production}
\usage{
calculate_urea_produced(df)
}
\arguments{
\item{df}{Dataframe with protein metabolism columns}
}
\value{
Numeric vector of urea production values
}
\description{
Calculate urea production
}
\keyword{internal}
