% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_water_heat_loss}
\alias{calculate_water_heat_loss}
\title{Calculate water used for heat loss}
\usage{
calculate_water_heat_loss(df)
}
\arguments{
\item{df}{Dataframe with water turnover and loss columns}
}
\value{
Numeric vector of water heat loss values
}
\description{
Calculate water used for heat loss
}
\keyword{internal}
