% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Function.R
\name{calculate_water_in_food}
\alias{calculate_water_in_food}
\title{Calculate water in food (vapor)}
\usage{
calculate_water_in_food(df)
}
\arguments{
\item{df}{Data frame containing FreeH2Oinfood}
}
\value{
Numeric vector of water vapor from food
}
\description{
Calculate water in food (vapor)
}
\keyword{internal}
