% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_wv_co2}
\alias{calculate_wv_co2}
\title{Calculate water vapor from CO2}
\usage{
calculate_wv_co2(df)
}
\arguments{
\item{df}{Dataframe with O2, urea, and influx columns}
}
\value{
Numeric vector of CO2 water vapor values
}
\description{
Calculate water vapor from CO2
}
\keyword{internal}
