% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_wv_fecal}
\alias{calculate_wv_fecal}
\title{Calculate water vapor from fecal loss}
\usage{
calculate_wv_fecal(df)
}
\arguments{
\item{df}{Dataframe with FecalH20Loss column}
}
\value{
Numeric vector of fecal water vapor values
}
\description{
Calculate water vapor from fecal loss
}
\keyword{internal}
