% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_wv_mouth}
\alias{calculate_wv_mouth}
\title{Calculate water vapor from mouth}
\usage{
calculate_wv_mouth(df)
}
\arguments{
\item{df}{Dataframe with H2OOral and Panting columns}
}
\value{
Numeric vector of mouth water vapor values
}
\description{
Calculate water vapor from mouth
}
\keyword{internal}
