% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{calculate_wv_sweating}
\alias{calculate_wv_sweating}
\title{Calculate water vapor from sweating}
\usage{
calculate_wv_sweating(df)
}
\arguments{
\item{df}{Dataframe with WaterHeatLoss column}
}
\value{
Numeric vector of sweat water vapor values
}
\description{
Calculate water vapor from sweating
}
\keyword{internal}
