% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_inputs}
\alias{combine_inputs}
\title{Create combinations dataframe from Species, Food, and Environment inputs}
\usage{
combine_inputs(species, food, environment)
}
\arguments{
\item{species}{Data frame with species parameters}

\item{food}{Data frame with food composition data}

\item{environment}{Data frame with environmental conditions}
}
\value{
Data frame with all input combinations
}
\description{
Create combinations dataframe from Species, Food, and Environment inputs
}
\keyword{internal}
