% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Function.R
\name{prepare_outputs_dataframe}
\alias{prepare_outputs_dataframe}
\title{Prepare outputs dataframe with initialized calculation columns}
\usage{
prepare_outputs_dataframe(inputs)
}
\arguments{
\item{inputs}{A dataframe containing inputs for calculations}
}
\value{
A dataframe with original inputs and new columns. Initialized to 0
}
\description{
Prepare outputs dataframe with initialized calculation columns
}
\keyword{internal}
